feat(postgres): add EventSubscriber implementation using LISTEN
