+++ eventually-postgres/Cargo.toml
@@ -21,6 +21,4 @@ tokio-postgres = { version = "0.5", features = ["with-serde_json-1"] }
 thiserror = "1.0"
 refinery = { version = "0.3.0", features = ["tokio-postgres"] }
 anyhow = "1.0.32"
-
-[dev-dependencies]
 tokio = { version = "0.2", features = ["sync"] }
